The prompt fold gets its own module; mobile stops bundling the agent processor
Review feedback on #2398 (bugbot + Misha): importing replayLlmRequest
pulled agent-processor-implementation — the StreamProcessor class and
Workers AI transport — into the mobile chat bundle. The pure fold
(reduce chain, projected context, request-body builder, message
renderers, ~560 lines) moves verbatim to agent-prompt-fold.ts; the
processor class, the replay lib, and the tests import from it. The
implementation keeps everything transport- and lifecycle-shaped.

Also drops the llmRequestOffset payload cast in the mobile card —
StreamEvent.payload is already Record<string, unknown>, so plain
optional access typechecks without assertion.
